chore: remove old k8 tests, rename _isolated replacements
